# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminImageOut": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "default_user": {
            "title": "Default User",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "min_disk_gb": {
            "title": "Min Disk Gb",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "os_family": {
            "title": "Os Family",
            "type": "string"
          },
          "templates": {
            "items": {
              "$ref": "#/components/schemas/TemplateRefOut"
            },
            "title": "Templates",
            "type": "array"
          },
          "tenant_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "visibility": {
            "title": "Visibility",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "os_family",
          "default_user",
          "min_disk_gb",
          "visibility",
          "tenant_id",
          "active",
          "created_at",
          "templates"
        ],
        "title": "AdminImageOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImageCreate": {
        "additionalProperties": false,
        "properties": {
          "cluster_id": {
            "format": "uuid",
            "title": "Cluster Id",
            "type": "string"
          },
          "default_user": {
            "default": "debian",
            "pattern": "^[a-z_][a-z0-9_-]{0,31}$",
            "title": "Default User",
            "type": "string"
          },
          "description": {
            "default": "",
            "maxLength": 1000,
            "title": "Description",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "os_family": {
            "default": "linux",
            "enum": [
              "linux",
              "windows"
            ],
            "title": "Os Family",
            "type": "string"
          },
          "template_vmid": {
            "minimum": 100.0,
            "title": "Template Vmid",
            "type": "integer"
          },
          "tenant_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "visibility": {
            "default": "public",
            "enum": [
              "public",
              "tenant"
            ],
            "title": "Visibility",
            "type": "string"
          }
        },
        "required": [
          "cluster_id",
          "template_vmid",
          "name"
        ],
        "title": "ImageCreate",
        "type": "object"
      },
      "ImageOut": {
        "description": "Tenant view: no servers, templates or provider identifiers.",
        "properties": {
          "default_user": {
            "title": "Default User",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "min_disk_gb": {
            "title": "Min Disk Gb",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "os_family": {
            "title": "Os Family",
            "type": "string"
          },
          "visibility": {
            "title": "Visibility",
            "type": "string"
          },
          "zone_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Zone Ids",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "os_family",
          "default_user",
          "min_disk_gb",
          "visibility",
          "zone_ids"
        ],
        "title": "ImageOut",
        "type": "object"
      },
      "ImageUpdate": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "default_user": {
            "anyOf": [
              {
                "pattern": "^[a-z_][a-z0-9_-]{0,31}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default User"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "ImageUpdate",
        "type": "object"
      },
...
      "TemplateAdd": {
        "additionalProperties": false,
        "properties": {
          "cluster_id": {
            "format": "uuid",
            "title": "Cluster Id",
            "type": "string"
          },
          "template_vmid": {
            "minimum": 100.0,
            "title": "Template Vmid",
            "type": "integer"
          }
        },
        "required": [
          "cluster_id",
          "template_vmid"
        ],
        "title": "TemplateAdd",
        "type": "object"
      },
      "TemplateOut": {
        "properties": {
          "disk_gb": {
            "title": "Disk Gb",
            "type": "integer"
          },
          "image_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Id"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "node": {
            "title": "Node",
            "type": "string"
          },
          "vmid": {
            "title": "Vmid",
            "type": "integer"
          }
        },
        "required": [
          "vmid",
          "name",
          "node",
          "disk_gb",
          "image_id"
        ],
        "title": "TemplateOut",
        "type": "object"
      },
      "TemplateRefOut": {
        "properties": {
          "cluster_id": {
            "format": "uuid",
            "title": "Cluster Id",
            "type": "string"
          },
          "cluster_name": {
            "title": "Cluster Name",
            "type": "string"
          },
          "disk_gb": {
            "title": "Disk Gb",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "template_vmid": {
            "title": "Template Vmid",
            "type": "integer"
          },
          "zone_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone Name"
          }
        },
        "required": [
          "id",
          "cluster_id",
          "cluster_name",
          "zone_name",
          "template_vmid",
          "disk_gb"
        ],
        "title": "TemplateRefOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Tenant-scoped routes (projects, instances, jobs, role bindings, dashboard) require the header `X-Tenant-Id: <tenant uuid>`; it is validated against the caller's memberships. It is omitted from each operation so generated clients can inject it once.",
    "title": "Cloud Manager API",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/clusters/{cluster_id}/templates": {
      "get": {
        "operationId": "list_templates_api_v1_admin_clusters__cluster_id__templates_get",
        "parameters": [
          {
            "in": "path",
            "name": "cluster_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Cluster Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TemplateOut"
                  },
                  "title": "Response List Templates Api V1 Admin Clusters  Cluster Id  Templates Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Templates",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/admin/images": {
      "get": {
        "operationId": "admin_list_images_api_v1_admin_images_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminImageOut"
                  },
                  "title": "Response Admin List Images Api V1 Admin Images Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Admin List Images",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "create_image_api_v1_admin_images_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminImageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Image",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/images/{image_id}": {
      "patch": {
        "operationId": "update_image_api_v1_admin_images__image_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Image Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminImageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Image",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/images/{image_id}/templates": {
      "post": {
        "operationId": "add_template_api_v1_admin_images__image_id__templates_post",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Image Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateAdd"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminImageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Template",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/images/{image_id}/templates/{template_id}": {
      "delete": {
        "operationId": "remove_template_api_v1_admin_images__image_id__templates__template_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Image Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "template_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Template Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminImageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Template",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/images": {
      "get": {
        "operationId": "list_images_api_v1_images_get",
        "parameters": [
          {
            "in": "query",
            "name": "zone_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Zone Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ImageOut"
                  },
                  "title": "Response List Images Api V1 Images Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Images",
        "tags": [
          "images"
        ]
      }
    },
...
  }
}